refactor:폴더 명 수정
